# api/openapi.yaml
openapi: 3.1.0
info:
  title: Dira Core API
  version: 0.1.0
  description: >-
    Socle de la plateforme Dira : identité, portefeuilles, paiements,
    notifications, véhicules et notes. Les verticales `dira-food-api`
    (livraison) et `dira-vtc-api` (courses) se posent dessus.


    ⚠️ Derrière la passerelle, ces routes sont servies à la RACINE
    (`/api/v1/...`), sans préfixe de verticale : l'authentification et le
    portefeuille sont les mêmes pour les deux, et les préfixer obligerait une
    application à choisir une verticale juste pour se connecter.


    ⚠️ Les routes `/internal/...` NE FIGURENT PAS ici. Elles portent les
    pouvoirs d'un service — débiter un portefeuille, ouvrir un compte — sont
    gardées par un secret partagé, et ne sont appelées par aucune application.
    Leur liste vit dans `internal/serviceapi/`, et un test refuse toute route
    qui s'y ajouterait sans être déclarée.


    **Erreurs.** Une seule enveloppe : `{"error":{"code","message"}}`, le
    message dans la langue de `Accept-Language`. Un `validation_failed`
    (422) ajoute `fields` — les clés JSON en cause — et, quand la valeur
    d'un champ n'est pas en cause, `reason` : `unknown_field` (une clé que
    la route ne connaît pas : elle est REFUSÉE, pas ignorée, pour qu'une
    application ne croie pas avoir enregistré ce qui a été jeté) ou
    `invalid_json`.


    **Téléphones.** Toujours en E.164 avec le `+` et l'indicatif ; les
...


...
    `POST /me/country/resolve`.
servers:
...
paths:
...
  /admin/finance/billing:
    get:
      tags: [finance]
      summary: La facturation du pays courant
      description: >-
        Par métier (`rides`, `deliveries`) : `agent_charge` ∈ tokens |
        commission, `commission_pct`, la grille de jetons (`tokens_per_ride`,
        `tokens_per_km`, `tokens_min`, `tokens_max` — la livraison garde sa
        grille `platform_pricing` quand ils sont à zéro), `client_charge_at` ∈
        request | accept | start | complete (portefeuille Dira seulement —
        le mobile money se paie d'avance, les espèces à la remise),
        `debt_limit_xof`. Défauts : courses en commission à la classe, client
        débité à l'acceptation ; livraison en jetons, client débité à la
        commande. Les verticales le lisent (`GET /internal/finance/billing`).
      responses:
        "200":
          description: Billing
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Billing"}
    put:
      tags: [finance]
      summary: Régler la facturation du pays courant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rides: {$ref: "#/components/schemas/VerticalBilling"}
                deliveries: {$ref: "#/components/schemas/VerticalBilling"}
      responses:
        "200": {description: Billing}
        "422": {description: "`invalid billing` — `reason` dit quoi"}
  /admin/finance/overview:
    get:
      tags: [finance]
      summary: L'aperçu comptable d'une période
      description: >-
        La balance des comptes du plan (débit, crédit, solde — un passif est
        ce que la plateforme doit, un produit ce qu'elle a gagné), les
        indicateurs (`kpis` : produits par nature, charges, net, passifs,
        créances, transit, à classer, encaissé), le détail par motif et par
        jour, l'état des portefeuilles à l'instant (`wallets`), le nombre de
        constats ouverts et le dernier balayage. `balanced` dit si la période
        s'équilibre. Défaut : le dernier mois.
      parameters:
        - {name: from, in: query, schema: {type: string, format: date}}
        - {name: to, in: query, schema: {type: string, format: date}}
      responses:
        "200": {description: Overview}
  /admin/finance/journal:
    get:
      tags: [finance]
      summary: Le journal — une écriture par mouvement, en partie double
      parameters:
        - {name: account, in: query, schema: {type: string}, description: un compte du plan (`GET /admin/finance/accounts`)}
        - {name: reason, in: query, schema: {type: string}}
        - {name: service, in: query, schema: {type: string, enum: [core, vtc, food]}}
        - {name: ref_kind, in: query, schema: {type: string}}
        - {name: ref_id, in: query, schema: {type: string}}
        - {name: owner_id, in: query, schema: {type: string}}
        - {name: unit, in: query, schema: {type: string, enum: [xof, token]}}
        - {name: from, in: query, schema: {type: string, format: date}}
        - {name: to, in: query, schema: {type: string, format: date}}
        - {name: cursor, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, maximum: 200}}
      responses:
        "200":
          description: Entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/AccountingEntry"}}
                  next_cursor: {type: string}
...
  /admin/finance/findings:
    get:
      tags: [finance]
      summary: Les constats du balayage d'intégrité
      description: >-
        `wallet_drift` (un solde stocké ≠ la somme de ses mouvements — jetons,
        argent, promo ou dette), `negative_balance`, `missing_entry` (un
        mouvement sans écriture), `entry_unbalanced`, `trial_unbalanced` (la
        balance générale). Dédoublonnés par empreinte, `occurrences` et
        `last_seen_at` disent depuis quand ; un écart disparu passe
        `resolved` au balayage suivant. `status` ∈ open | acknowledged |
        resolved (défaut : non résolus).
      parameters: [{name: status, in: query, schema: {type: string}}]
      responses:
        "200": {description: Findings, avec `last_run`}
  /admin/finance/findings/{id}/ack:
    post:
      tags: [finance]
      summary: Prendre acte d'un constat (avec une note)
      parameters: [{name: id, in: path, required: true, schema: {type: string}}]
      requestBody:
        content:
          application/json:
            schema: {type: object, properties: {note: {type: string, maxLength: 500}}}
      responses:
        "200": {description: Finding}
...
components:
  schemas:
...
    VerticalBilling:
      type: object
      properties:
        agent_charge: {type: string, enum: [tokens, commission]}
        commission_pct: {type: integer, minimum: 0, maximum: 100}
        tokens_per_ride: {type: integer}
        tokens_per_km: {type: integer}
        tokens_min: {type: integer}
        tokens_max: {type: integer}
        client_charge_at: {type: string, enum: [request, accept, start, complete]}
        debt_limit_xof: {type: integer}
    Billing:
      type: object
      properties:
        country: {type: string}
        rides: {$ref: "#/components/schemas/VerticalBilling"}
        deliveries: {$ref: "#/components/schemas/VerticalBilling"}
        updated_by: {type: string}
        updated_at: {type: string, format: date-time}
    AccountingEntry:
      type: object
      properties:
        id: {type: string}
        country: {type: string}
        at: {type: string, format: date-time}
        service: {type: string, enum: [core, vtc, food]}
        reason: {type: string}
        ref_kind: {type: string}
        ref_id: {type: string}
        transaction_id: {type: string}
        wallet_id: {type: string}
        owner_id: {type: string}
        unit: {type: string, enum: [xof, token]}
        amount: {type: integer, description: en francs — les jetons sont valorisés au prix du jeton}
        tokens: {type: integer}
        description: {type: string}
        lines:
          type: array
          items:
            type: object
            properties:
              account: {type: string}
              debit: {type: integer}
              credit: {type: integer}
...
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
security:
  - bearerAuth: []